+++ include/Gaffer/TypeIds.h
@@ -155,6 +155,7 @@ enum TypeId
 	Box2fVectorDataPlugTypeId = 110109,
 	PatternMatchTypeId = 110110,
 	Int64VectorDataPlugTypeId = 110111,
+	ObjectMatrixTypeId = 110112,
 
 	LastTypeId = 110159,